Move Slack user interaction to a separate file
