+++ door-open-service/node-backend/src/httpInterface.js
